邀请报名取消的时候不存在订单的异常处理

+++ api/bll/order.py
@@ -43,9 +43,10 @@ def applyParent(request):
     #查找教师
     teachers = user.teacher_set.all()
     #查找家长订单
-    pd = ParentOrder.objects.filter(pd_id=pd_id)
-    if len(pd) and len(teachers):
+    pds = ParentOrder.objects.filter(pd_id=pd_id)
+    if len(pds) and len(teachers):
         teacher = teachers[0]
+        pd = pds[0]
         if method == 1:
             #老师报名家长
             #老师可以报名多个家长!!
@@ -53,13 +54,13 @@ def applyParent(request):
             try:
                 with transaction.atomic():
                     #新建订单
-                    order = OrderApply(apply_type=1, pd=pd[0],tea=teacher,parent_willing=1,teacher_willing=2,
+                    order = OrderApply(apply_type=1, pd=pd,tea=teacher,parent_willing=1,teacher_willing=2,
                                        pass_not=1,update_time=timezone.now(),expectation=expectation,finished=0)
                     order.save()
                     message_title = teacher.name + u"向您报名!"
                     message_content = teacher.name + u"向您报名!请到“我的老师”处查看详细信息!"
                     #新建消息
-                    message = Message(sender=user, receiver=pd[0].wechat,message_title=message_title, message_content=message_content,status=0)
+                    message = Message(sender=user, receiver=pd.wechat,message_title=message_title, message_content=message_content,status=0)
                     message.save()
                     #TODO:推送到微信端
 
@@ -72,16 +73,18 @@ def applyParent(request):
             try:
                 with transaction.atomic():
                     #删除订单
-                    order = OrderApply.objects.get(apply_type=1, pd=pd[0],tea=teacher)
-                    # order.finished = 1
-                    # order.save()
-                    order.delete()
-                    message_title = teacher.name + u"取消了报名!"
-                    message_content = teacher.name + u"取消了报名!"
-                    #新建消息
-                    message = Message(sender=user, receiver=teacher.wechat, message_title=message_title, message_content=message_content,status=0)
-                    message.save()
-                    #TODO:推送到微信端
+                    orders = OrderApply.objects.filter(apply_type=1, pd=pd,tea=teacher)
+                    if len(orders):
+                        order = orders[0]
+                        order.delete()
+                        message_title = teacher.name + u"取消了报名!"
+                        message_content = teacher.name + u"取消了报名!"
+                        #新建消息
+                        message = Message(sender=user, receiver=teacher.wechat, message_title=message_title, message_content=message_content,status=0)
+                        message.save()
+                    else:
+                        return JsonError(u"找不到该订单")
+                        #TODO:推送到微信端
 
             except Exception,e:
                 return JsonError(e.message)
@@ -107,11 +110,8 @@ def inviteTeacher(request):
     tea_id = int(request.data.get("tea_id", -1))
     method = int(request.data.get("type", -1))
     user = AuthUser.objects.get(username=request.user.username)
-    #查找家长订单
-    #TODO： 每个家长是否只有一个需求
-    parentorders = user.parentorder_set.all()
-    #查找教师
-    teachers = Teacher.objects.filter(tea_id=tea_id)
+    parentorders = user.parentorder_set.all()   #查找家长订单
+    teachers = Teacher.objects.filter(tea_id=tea_id)    #查找教师
     if len(parentorders) and len(teachers):
         parentorder = parentorders[0]
         teacher = teachers[0]
@@ -120,7 +120,7 @@ def inviteTeacher(request):
             #如果家长已经邀请了老师，返回错误
             oa = OrderApply.objects.filter(apply_type=2, pd=parentorder)
             if len(oa) != 0:
-                return JsonError("已经邀请了老师")
+                return JsonError(u"已经邀请了老师!")
             #事务
             try:
                 with transaction.atomic():
@@ -142,17 +142,18 @@ def inviteTeacher(request):
             try:
                 with transaction.atomic():
                     #取消订单
-                    order = OrderApply.objects.get(apply_type=2, pd=parentorder,tea=teacher)
-                    order.delete()
-                    # order.finished = 1
-                    # order.save()
-                    message_title = parentorder.name + u"取消了邀请!"
-                    message_content = parentorder.name + u"取消了邀请!"
-                    #新建消息
-                    message = Message(sender=user, receiver=teacher.wechat, message_title=message_title, message_content=message_content,status=0)
-                    message.save()
-                    #TODO:推送到微信端
-
+                    orders = OrderApply.objects.filter(apply_type=2, pd=parentorder,tea=teacher)
+                    if len(orders):
+                        order = orders[0]
+                        order.delete()
+                        message_title = parentorder.name + u"取消了邀请!"
+                        message_content = parentorder.name + u"取消了邀请!"
+                        #新建消息
+                        message = Message(sender=user, receiver=teacher.wechat, message_title=message_title, message_content=message_content,status=0)
+                        message.save()
+                        #TODO:推送到微信端
+                    else:
+                        return JsonError(u"找不到订单！")
             except Exception,e:
                 return JsonError(e.message)
             return JsonResponse()
